add currency function file
contains  currency_detection function which is responsible for returning of the label of the currency found in the photo using tensorflow-lite 
In this module i used TensorFlow Lite to be able to run our trained model on the raspberry pi, to be able to perform an inference we must run it through an interpreter. ( Inference is the process of executing TensorFlow lite model on-device).